+++ src/learned_koopman/models/__init__.py
@@ -1,5 +1,11 @@
 from learned_koopman.models.baselines import ResidualMLP
 from learned_koopman.models.energy_conditioned import EnergyConditionedRotation
 from learned_koopman.models.fixed_koopman import FixedKoopmanAE
+from learned_koopman.models.separatrix_atlas import SeparatrixAtlas
 
-__all__ = ["EnergyConditionedRotation", "FixedKoopmanAE", "ResidualMLP"]
+__all__ = [
+    "EnergyConditionedRotation",
+    "FixedKoopmanAE",
+    "ResidualMLP",
+    "SeparatrixAtlas",
+]